# docs/swagger.yaml
definitions:
...
  common.BaseResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  project.CreateProjectRequest:
    properties:
      description:
        maxLength: 1000
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - name
    type: object
  project.ProjectListResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      projects:
        items:
          $ref: '#/definitions/project.ProjectResponse'
        type: array
      total:
        type: integer
    type: object
  project.ProjectResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  project.UpdateProjectRequest:
    properties:
      description:
        maxLength: 1000
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/projects:
    get:
      consumes:
      - application/json
      description: Get all projects with pagination
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Projects retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/common.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/project.ProjectListResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all projects
      tags:
      - Project
    post:
      consumes:
      - application/json
      description: Create a new project (only managers and admins)
      parameters:
      - description: Project creation request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/project.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Project created successfully
          schema:
            allOf:
            - $ref: '#/definitions/common.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/project.ProjectResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "409":
          description: Project name already exists
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new project
      tags:
      - Project
  /api/v1/projects/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing project and all its tasks
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Project deleted successfully
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete project
      tags:
      - Project
    get:
      consumes:
      - application/json
      description: Get a project by its ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Project retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/common.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/project.ProjectResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Get project by ID
      tags:
      - Project
    put:
      consumes:
      - application/json
      description: Update an existing project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Project update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/project.UpdateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Project updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/common.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/project.ProjectResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "409":
          description: Project name already exists
          schema:
            $ref: '#/definitions/common.BaseResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.BaseResponse'
      security:
      - ApiKeyAuth: []
      summary: Update project
      tags:
      - Project
...
swagger: "2.0"